requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloud-sql-python-connector==1.2.4
cryptography==41.0.1
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
frozenlist==1.3.3
google-auth==2.21.0
google-cloud==0.34.0
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
neo4j==5.10.0
numpy==1.25.0
pandas==2.0.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.17
typing-extensions==4.7.1
tzdata==2023.3
unicode==2.9
Unidecode==1.3.6
urllib3==1.26.16
validators==0.20.0
Werkzeug==2.3.6
WTForms==3.0.1
yarl==1.9.2
zipp==3.15.0
